refactor: controller & dto 수정
