+++ src/main/java/wooteco/subway/admin/dto/StationCreateRequest.java
@@ -3,7 +3,10 @@ package wooteco.subway.admin.dto;
 
 import wooteco.subway.admin.domain.Station;
 
+import javax.validation.constraints.NotBlank;
+
 public class StationCreateRequest {
+    @NotBlank
     private String name;
 
     public String getName() {
