+++ src/main/java/wooteco/subway/admin/service/StationService.java
@@ -2,6 +2,7 @@ package wooteco.subway.admin.service;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 import wooteco.subway.admin.domain.Station;
 import wooteco.subway.admin.dto.StationResponse;
 import wooteco.subway.admin.repository.StationRepository;
@@ -9,6 +10,7 @@ import wooteco.subway.admin.repository.StationRepository;
 import java.util.List;
 
 @Service
+@Transactional
 public class StationService {
     private final StationRepository stationRepository;
